Clean-up misplaced tests in `core_integration`
